Notify builder of auth changes in default toolbar
